Sell buildings that have lost their reqs (other than tech and improvement) when city is processed.
As a generalization of selling landlocked coastal buildings, add city_remove_invalid_buildings()
to check all reqs except Tech for each building during update_city_activities().

Buildings can already require e.g. a particular Terrain in City range, or a particular Government,
and this allows them to be automatically sold when those requirements disappear, as is done for
coastal buildings.

Exception for tech and improvement reqs is to keep existing behaviour in most cases.
(e.g. don't sell University if the University tech is lost somehow, or if the Library disappears.)

Further development idea: split the reqs vector in two, one for build-time checks, one for
every-turn checks and/or make this otherwise configurable.