Modify nukes to destroy city improvements too, and damage units randomly.
City improvements other than wonders and those that are resistant to sabotage
get destroyed with a 50 % chance each. Units get 10 to 40 points of damage
randomly and are destroyed if that is enough to bring them to 0 HP.

The numbers are totally pulled out of a hat and might require at least some
configurability.

Also call send_city_info() on nuked cities to make sure the new size
and visible imrovements are updated.