Client: pick up small wonders from the map, don't trust player_info fully
To go with a change to the server to not send known but unseen or any
foreign small wonders, pick them up from the city_info packets, which
contain remembered information. Hence, assume that if player_info says
the wonder isn't built, it may be the server just isn't telling.